requirements.txt
flask
pandas
openpyxl
psycopg2
python-dotenv
snowflake-connector-python
snowflake-connector-python[pandas]